+++ service/controller/resource/monitoring/prometheus/resource.go
@@ -2,7 +2,6 @@ package prometheus
 
 import (
 	"fmt"
-	"math"
 	"net/url"
 
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
@@ -18,7 +17,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/nodecounter"
+	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/prometheus"
 	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/pvcresizing"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/generic"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/key"
@@ -145,21 +144,9 @@ func toPrometheus(ctx context.Context, v interface{}, config Config) (metav1.Obj
 
 	labels[key.MonitoringLabel] = "true"
 
-	prometheusMemory := key.PrometheusDefaultMemory()
-	if !key.IsManagementCluster(config.Installation, cluster) || key.IsCAPIManagementCluster(config.Provider) {
-		nodeCount, err := nodecounter.CountClusterNodes(ctx, config.K8sClient, cluster)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-		if nodeCount > 2 {
-			// We consider that a node requires 500Mb of Prometheus RAM
-			prometheusMemory = resource.NewQuantity(
-				int64(math.Floor(
-					1024*1024*512*float64(nodeCount),
-				)),
-				resource.DecimalSI,
-			)
-		}
+	prometheusMemory, err := prometheus.ComputePrometheusMinMemory(ctx, config.K8sClient, cluster, config.Installation, config.Provider)
+	if err != nil {
+		return nil, microerror.Mask(err)
 	}
 
 	image := fmt.Sprintf("%s/%s:%s", config.Registry, config.ImageRepository, config.Version)
@@ -174,9 +161,9 @@ func toPrometheus(ctx context.Context, v interface{}, config Config) (metav1.Obj
 			// This forces us to use the static config defined in resource/alerting/alertmanagerwiring.
 			AdditionalAlertManagerConfigs: &corev1.SecretKeySelector{
 				LocalObjectReference: corev1.LocalObjectReference{
-					Name: key.AlertManagerSecretName(),
+					Name: key.AlertmanagerSecretName(),
 				},
-				Key: key.AlertManagerKey(),
+				Key: key.AlertmanagerKey(),
 			},
 
 			CommonPrometheusFields: promv1.CommonPrometheusFields{
@@ -227,6 +214,8 @@ func toPrometheus(ctx context.Context, v interface{}, config Config) (metav1.Obj
 						corev1.ResourceMemory: *prometheusMemory,
 					},
 					Limits: corev1.ResourceList{
+						// cpu: 150m
+						corev1.ResourceCPU:    *key.PrometheusDefaultCPULimit(),
 						corev1.ResourceMemory: *prometheusMemory,
 					},
 				},