+++ service/key/key.go
@@ -2,6 +2,7 @@ package key
 
 import (
 	"fmt"
+	"math"
 
 	"github.com/giantswarm/microerror"
 	v1 "k8s.io/api/core/v1"
@@ -20,11 +21,14 @@ const (
 	DefaultServicePriority string = "highest"
 	DefaultOrganization    string = "giantswarm"
 
-	ClusterLabel         string = "giantswarm.io/cluster"
-	MonitoringLabel      string = "giantswarm.io/monitoring"
-	OrganizationLabel    string = "giantswarm.io/organization"
-	ServicePriorityLabel string = "giantswarm.io/service-priority"
-	TeamLabel            string = "application.giantswarm.io/team"
+	ClusterLabel                 string = "giantswarm.io/cluster"
+	MonitoringLabel              string = "giantswarm.io/monitoring"
+	OrganizationLabel            string = "giantswarm.io/organization"
+	ServicePriorityLabel         string = "giantswarm.io/service-priority"
+	TeamLabel                    string = "application.giantswarm.io/team"
+	OpsGenieApiKey               string = "opsGenieApiKey" // #nosec G101
+	AlertmanagerGlobalSecretName string = "alertmanager-global"
+
 	// PrometheusCPULimitCoefficient is the number used to compute the CPU limit from the CPU request.
 	// It is used when computing VPA settings, to set `max requests` so that `max limits` respects MaxCPU factor.
 	PrometheusCPULimitCoefficient float64 = 1.5
@@ -205,12 +209,13 @@ func PrometheusDefaultCPU() *resource.Quantity {
 	return resource.NewMilliQuantity(100, resource.DecimalSI)
 }
 
-func PrometheusDefaultMemory() *resource.Quantity {
-	return resource.NewQuantity(1024*1024*1024, resource.DecimalSI)
-}
-
-func AlertmanagerKey() string {
-	return "alertmanager.yaml"
+func PrometheusDefaultCPULimit() *resource.Quantity {
+	return resource.NewMilliQuantity(
+		int64(math.Floor(
+			100*PrometheusCPULimitCoefficient,
+		)),
+		resource.DecimalSI,
+	)
 }
 
 func PrometheusPort() int32 {
@@ -257,11 +262,11 @@ func PrometheusSTSName(cluster metav1.Object) string {
 	return fmt.Sprintf("prometheus-%s", ClusterID(cluster))
 }
 
-func AlertManagerSecretName() string {
+func AlertmanagerSecretName() string {
 	return "alertmanager-config"
 }
 
-func AlertManagerKey() string {
+func AlertmanagerKey() string {
 	return "alertmanager-additional.yaml"
 }
 
@@ -308,18 +313,6 @@ func CAFilePath() string {
 	return "/var/run/secrets/kubernetes.io/serviceaccount/ca.crt"
 }
 
-func AlertmanagerConfigMapName() string {
-	return "alertmanager"
-}
-
-func AlertmanagerConfigMapNamespace() string {
-	return monitoring
-}
-
-func AlertmanagerConfigMapKey() string {
-	return "config.yml"
-}
-
 // IsCAPICluster returns false if the cluster has any of the legacy labels such as azure-operator.giantswarm.io/version.
 func IsCAPICluster(obj metav1.Object) bool {
 	// TODO once we have migrated all clusters to CAPI, we can remove this
@@ -348,8 +341,4 @@ func IsCAPICluster(obj metav1.Object) bool {
 
 func IngressClassName() string {
 	return "nginx"
-}
-
-func OpsgenieKey() string {
-	return "opsgenie.key"
 }